1-cat.py added ,2- We added cat ,move ,remove ,ls ,mkdir commands functionality into cli.py ,3-and make some changes into copy.py , ls.py ,mkdir.py ,remove.py ,move.py ,4-And we created test_cli_cat.py file to test cat commend .

+++ src/fman/commands/remove.py
@@ -1,7 +1,7 @@
 import shutil
 from pathlib import Path 
 
-def remove(path:str,force:bool=False)->None:
+def removePath(path:str,force:bool=False)->None:
     p=Path(path)
 
     if not p.exists():
@@ -13,11 +13,10 @@ def remove(path:str,force:bool=False)->None:
 
             print("Delete operation cancelled.")
             return
-
-    if p.is_file():
-        p.unlink()
-    elif p.is_dir():
-        shutil.rmtree(p)
-
-    raise NotImplementedError("remove() is not yet implemented.")
         
+        if yes_or_no.lower() == "y":
+
+            if p.is_file():
+                p.unlink()
+            elif p.is_dir():
+                shutil.rmtree(p)
